requirements.txt
anyio==3.6.1
click==8.1.3
ecdsa==0.18.0
Faker==14.2.0
fastapi==0.82.0
h11==0.13.0
idna==3.3
pyasn1==0.4.8
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.19.1
typing_extensions==4.3.0
uvicorn==0.18.3
